[Bronze V] Title: 합, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/8393. 합/합.cc
@@ -4,9 +4,10 @@ using namespace std;
 
 int main()
 {
-    long n, ans;
+    int n, ans;
     cin >> n;
     ans = 0;
+
     for (int i = 0; i <= n; i++)
     {
         ans += i;
